--- Search/src/Movie.css
@@ -1,47 +1,0 @@
-.movies .movie {
-    width: 45%;
-    background-color: white;
-    margin-bottom: 70px;
-    display: flex;
-    align-items: flex-start;
-    justify-content: space-between;
-    font-weight: 300;
-    padding: 20px;
-    border-radius: 5px;
-    color: #b2b3bc;
-    box-shadow: 0 13px 27px -5px rgba(50, 50, 93, 0.25),
-      0 8px 16px -8px rgba(0, 0, 0, 0.3), 0 -6px 16px -6px rgba(0, 0, 0, 0.025);
-  }
-  
-.movie img {
-    position: relative;
-    top: -50px;
-    max-width: 150px;
-    width: 100%;
-    margin-right: 30px;
-    box-shadow: 0 30px 60px -12px rgba(50, 50, 93, 0.25),
-      0 8px 36px -18px rgba(0, 0, 0, 0.3), 0 -12px 36px -8px rgba(0, 0, 0, 0.025);
-  }
-.movie .movie__year,
-.movie .movie__title {
-  margin: 0;
-  font-weight: 300;
-}
-
-.movie .movie__title {
-    margin-bottom: 5px;
-    font-size: 24px;
-    color: #2c2c2c;
-  }
-  
-.movie .movie__genres {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    margin: 5px 0px;
-  }
-.movie .movie__year {
-    margin-right: 10px;
-    font-size: 14px;
-  }
