--- Search/src/index.js
@@ -1,13 +1,0 @@
-// import React from 'react';
-// import ReactDOM from 'react-dom';
-// import App from './App';
-
-// ReactDOM.render(<App />, document.getElementById('root'));
-
-
-import React from 'react';
-import * as ReactDOM from 'react-dom/client';
-import App from './App';
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(<App />);
