+++ src/routes/Search.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import axios from 'axios';
 import SearchMovie from '../components/SearchMovie';
-import styles from './Search.module.css';
-// import Detail from './Detail';
+import "./Home.css";
+import "./Search.css";
 
 class Search extends React.Component {
   state = {
@@ -13,15 +13,16 @@ class Search extends React.Component {
 
 
   getSearchMovie = async () => {
-    const ID_KEY = 'qiMNfCrKAD2YwpKZAnmG';
+    const ID_KEY = 'qiMNfCrKAD2YwpKZAnmG'; // 네이버 개발 API 발급받고 Key 입력!!
     const SECRET_KEY = 'X_9ZY5oH3v';
     const search = this.state.value;
     try {
       if (search === "") {
         this.setState({movies: [], isLoading: false})
       } else {
-        const {data: {items}} 
-          = await axios.get('/v1/search/movie.json',{
+        const {data: {
+            items
+          }} = await axios.get('/v1/search/movie.json',{
             params:{
               query: search,
               display: 20
@@ -38,7 +39,6 @@ class Search extends React.Component {
       console.log(error);
     }
   };
-
 
   componentDidMount() {
     this.getSearchMovie();
@@ -68,14 +68,14 @@ class Search extends React.Component {
           </div>)
           : (<form onSubmit={this.handleSubmit}>
             <div>
-              <div className={styles.input_div}>
-                {/* <h1>영화 검색</h1> */}
-                <input className={styles.input_search} type="text" value={this.state.value} onChange={this.handleChange} placeholder="영화를 검색해 보세요."/>
+              <div className="input_div">
+                <h1>영화 검색</h1>
+                <input className="input_search" type="text" value={this.state.value} onChange={this.handleChange} placeholder="영화를 검색해 보세요."/>
               </div>
-              <div className={styles.movies}>
+              <div className="movies">
                 {movies.map(movie => (
                 	<SearchMovie key={movie.link} id={movie.link} year={movie.pubDate} title={movie.title} poster={movie.image} rating={movie.userRating} director={movie.director} actor={movie.actor}
-					        />))
+					/>))
                  }
               </div>
             </div>
